requirements.txt
paho-mqtt>=1.6.1
attrs>=21.2.0